Added an ExecuteRaw function to obtain JSON data directly (Fixes #4).
All 94 tests pass.

+++ Solution/RexConnectClient.Core/Result/IResponseResult.cs
@@ -15,7 +15,7 @@ namespace RexConnectClient.Core.Result {
 		string ResponseJson { get; }
 
 		bool IsError { get; }
-		int ExecutionMilliseconds { get; set; }
+		double ExecutionMilliseconds { get; set; }
 
 
 		////////////////////////////////////////////////////////////////////////////////////////////////
